requirements.txt
Django

django-allauth

gunicorn

django-bootstrap-v5

dj_database_url

psycopg2

whitenoise

django-storages

boto3

Pillow

python-dotenv